+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 - Review the organization of the build targets
 - Use `displayName` on task functions, instead of separate `generateTaskName` functions.
 
+## Next
+
+- **[Internal]** Convert gulpfile to Typescript, use `ts-node` to run it.
+
 ## 0.14.3 (2017-07-16)
 
 - **[Feature]** Add `:lint:fix` project task to fix some lint errors.